+++ src/test_ocr.py
@@ -6,8 +6,27 @@ import os.path
 import logging
 import matplotlib.pyplot as plt
 
-from image_processing import plot_images, extract_chars, cv_skeletonize
 from ocr import OCR
+
+def plot_images(data, rows, cols, cmap='gray'):
+    if(len(data) > 0):
+        i = 0
+        for title, image in data.items():
+            #logging.debug(title)    
+            plt.subplot(rows,cols,i+1),plt.imshow(image,cmap)
+            plt.title(title)
+            plt.xticks([]),plt.yticks([])
+            i += 1
+        plt.show()
+
+def display_images(img_list, row, col):
+    if(len(img_list) > 0):
+        images = {}
+        n = 0
+        for img in img_list:
+            n += 1
+            images[str(n)] = img
+        plot_images(images, row, col, cmap='gray')
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Testing OCR.')
@@ -27,15 +46,9 @@ if __name__ == '__main__':
     if hasFrame:
         images = {}
         images['frame'] = frame
-        
-        characteres, img, mask = extract_chars(frame) 
-        images['img'] = img.copy()
-        img = cv_skeletonize(img)
-        img = np.uint8(img / 255)
-        images['skel'] = img.copy()
 
         ocr = OCR(model_filename="../config/attention_ocr_model.pth", use_cuda=False, threshold=0.7)
-        pred = ocr.predict(img)
+        pred = ocr.predict(frame)
         logging.info(f'Prediction: {pred}')
 
         plot_images(images, 1, 3, cmap='gray')
